Split up the ClientNetworking class into separate namespaces.
