+++ Source/Client/Networking/ClientNetListener.cs
@@ -1,0 +1,62 @@
+using System.Net;
+using System.Net.Sockets;
+using LiteNetLib;
+using Multiplayer.Common;
+using Verse;
+
+namespace Multiplayer.Client.Networking
+{
+    public class MpClientNetListener : INetEventListener
+    {
+        public void OnPeerConnected(NetPeer peer)
+        {
+            IMultiplayerConnection conn = new MpNetMultiplayerConnection(peer);
+            conn.username = Multiplayer.username;
+            conn.State = ConnectionStateEnum.ClientJoining;
+
+            Multiplayer.session.client = conn;
+            Multiplayer.session.ReapplyPrefs();
+
+            MpLog.Log("Net client connected");
+        }
+
+        public void OnNetworkError(IPEndPoint endPoint, SocketError error)
+        {
+            Log.Warning($"Net client error {error}");
+        }
+
+        public void OnNetworkReceive(NetPeer peer, NetPacketReader reader, DeliveryMethod method)
+        {
+            byte[] data = reader.GetRemainingBytes();
+            Multiplayer.HandleReceive(new ByteReader(data), method == DeliveryMethod.ReliableOrdered);
+        }
+
+        public void OnPeerDisconnected(NetPeer peer, DisconnectInfo info)
+        {
+            var reader = new ByteReader(info.AdditionalData.GetRemainingBytes());
+            Multiplayer.session.HandleDisconnectReason((MpDisconnectReason)reader.ReadByte(), reader.ReadPrefixedBytes());
+
+            ConnectionStatusListeners.TryNotifyAll_Disconnected();
+
+            OnMainThread.StopMultiplayer();
+            MpLog.Log("Net client disconnected");
+        }
+
+        private static string DisconnectReasonString(DisconnectReason reason)
+        {
+            switch (reason)
+            {
+                case DisconnectReason.ConnectionFailed: return "Connection failed";
+                case DisconnectReason.ConnectionRejected: return "Connection rejected";
+                case DisconnectReason.Timeout: return "Timed out";
+                case DisconnectReason.HostUnreachable: return "Host unreachable";
+                case DisconnectReason.InvalidProtocol: return "Invalid library protocol";
+                default: return "Disconnected";
+            }
+        }
+
+        public void OnConnectionRequest(ConnectionRequest request) { }
+        public void OnNetworkLatencyUpdate(NetPeer peer, int latency) { }
+        public void OnNetworkReceiveUnconnected(IPEndPoint remoteEndPoint, NetPacketReader reader, UnconnectedMessageType messageType) { }
+    }
+}
