+++ Source/Client/Persistent/CaravanForming.cs
@@ -1,14 +1,13 @@
-﻿using Harmony;
+﻿using System;
+using System.Collections.Generic;
+using Harmony;
+using Multiplayer.Client.Sync;
 using RimWorld;
 using RimWorld.Planet;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using UnityEngine;
 using Verse;
 
-namespace Multiplayer.Client
+namespace Multiplayer.Client.Persistent
 {
     public class CaravanFormingSession : IExposable, ISessionWithTransferables
     {
