+++ Source/Client/Persistent/Trading.cs
@@ -1,20 +1,16 @@
-﻿using Harmony;
-using Multiplayer.Common;
-using RimWorld;
-using RimWorld.Planet;
-using System;
-using System.Collections.Generic;
-using System.Diagnostics;
+﻿using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using System.Reflection.Emit;
-using System.Text;
-using UnityEngine;
+using Harmony;
+using Multiplayer.Client.Sync;
+using RimWorld;
+using RimWorld.Planet;
 using Verse;
 using Verse.AI;
 using Verse.AI.Group;
 
-namespace Multiplayer.Client
+namespace Multiplayer.Client.Persistent
 {
     public class MpTradeSession : IExposable, ISessionWithTransferables
     {
