+++ Source/Client/TickPatch.cs
@@ -8,6 +8,7 @@ using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
 using System.Threading;
+using Multiplayer.Client.Comp;
 using UnityEngine;
 using Verse;
