+++ Source/Multiplayer.csproj
@@ -78,6 +78,10 @@
     <Compile Include="Client\Desyncs\SyncCoordinator.cs" />
     <Compile Include="Client\Desyncs\UserReadableDesyncInfo.cs" />
     <Compile Include="Client\ModPatches.cs" />
+    <Compile Include="Client\Networking\LocalClientToServerConnection.cs" />
+    <Compile Include="Client\Networking\ClientUtil.cs" />
+    <Compile Include="Client\Networking\ClientNetListener.cs" />
+    <Compile Include="Client\Networking\SteamClientToServerConnection.cs" />
     <Compile Include="Client\Optimizations.cs" />
     <Compile Include="Client\SettingsPatches.cs" />
     <Compile Include="Client\TickPatch.cs" />
@@ -87,7 +91,6 @@
     <Compile Include="Client\Persistent\TransporterLoading.cs" />
     <Compile Include="Client\Windows\ChatWindow.cs" />
     <Compile Include="Client\ClientConnection.cs" />
-    <Compile Include="Client\ClientNetworking.cs" />
     <Compile Include="Client\Windows\ConnectingWindow.cs" />
     <Compile Include="Client\CrossRefs.cs" />
     <Compile Include="Client\DebugTools.cs" />
@@ -133,10 +136,13 @@
     <Compile Include="Common\MpLog.cs" />
     <Compile Include="Common\MultiplayerServer.cs" />
     <Compile Include="Common\Networking.cs" />
+    <Compile Include="Common\Networking\SteamBaseConnection.cs" />
     <Compile Include="Common\Packets.cs" />
     <Compile Include="Common\ServerConnection.cs" />
     <Compile Include="Common\Version.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="Server\Networking\LocalServerToClientConnection.cs" />
+    <Compile Include="Server\Networking\SteamServerToClientConnection.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
